+++ qiskit_ibm_runtime/program/result_decoder.py
@@ -15,7 +15,7 @@
 import json
 from typing import Any
 
-from qiskit_ibm_runtime.utils.runtime import RuntimeDecoder
+from qiskit_ibm_runtime.utils import RuntimeDecoder
 
 
 class ResultDecoder:
